+++ test/test.rs
@@ -23,7 +23,7 @@ mod test_mq;
 mod test_net;
 mod test_nix_path;
 mod test_poll;
-#[cfg(not(target_os = "redox"))]
+#[cfg(not(any(target_os = "redox", target_os = "fuchsia")))]
 mod test_pty;
 #[cfg(any(target_os = "android",
           target_os = "linux"))]