+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added `mremap` (#[1306](https://github.com/nix-rust/nix/pull/1306))
 
 - Added `personality` (#[1331](https://github.com/nix-rust/nix/pull/1331))
+
+- Added limited Fuchsia support (#[1285](https://github.com/nix-rust/nix/pull/1285))
+
 ### Fixed
 ### Changed